# api/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "SendIT is a courier service that helps users deliver parcels to different destinations.",
    "version": 1,
    "title": "SendIT API Documentation",
    "contact": {
      "name": "Emmanuel Columbus Mugenyi",
      "email": "ecmugenyi@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "127.0.0.1:5000",
  "basePath": "/api/v1/docs",
  "tags": [
    {
      "name": "user",
      "description": "Operations about user",
      "externalDocs": {
        "description": "Find out more about our store",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "parcel",
      "description": "Everything about Parcel delivery orders",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Login a user",
        "description": "",
        "operationId": "loginUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Login user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create a user account.",
        "description": "",
        "operationId": "createUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserSignUp"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/parcels": {
      "post": {
        "tags": [
          "parcel"
        ],
        "summary": "Create a parcel delivery order",
        "description": "",
        "operationId": "createParcel",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Authorization Token",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Parcel object that needs to be created",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateParcel"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      },
      "get": {
        "tags": [
          "parcel"
        ],
        "summary": "Fetch all parcel delivery orders",
        "description": "",
        "operationId": "fetchParcels",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      }
    },
    "/parcels/{parcelId}": {
      "get": {
        "tags": [
          "parcel"
        ],
        "summary": "Fetch a specific delivery order",
        "description": "",
        "operationId": "findSpecificParcel",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          }
        },
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Token",
            "required": false,
            "type": "string"
          },
          {
            "name": "parcelId",
            "in": "path",
            "description": "Fetch parcel py Id",
            "required": true,
            "type": "integer"
          }
        ],
        "security": [
          {
            "token": []
          }
        ]
      }
    },
    "/users/{userId}/parcels": {
      "get": {
        "tags": [
          "parcel"
        ],
        "summary": "Find Parcels by user",
        "description": "Fetch all parcel delivery order by a specific user.  ",
        "operationId": "findParcelByUser",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          }
        },
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string",
            "description": "Authorization Token"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "security": [
          {
            "token": []
          }
        ]
      }
    },
    "/parcels/{parcelId}/cancel": {
      "patch": {
        "tags": [
          "parcel"
        ],
        "summary": "Cancel a specific order",
        "description": "",
        "operationId": "cancelOrder",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          }
        },
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string",
            "description": "Authorization Token"
          },
          {
            "name": "parcelId",
            "in": "path",
            "description": "Parcel id to cancel",
            "required": true,
            "type": "integer",
            "format": "integer"
          }
        ],
        "security": [
          {
            "token": []
          }
        ]
      }
    },
    "/parcels/{parcelId}/destination": {
      "patch": {
        "tags": [
          "parcel"
        ],
        "summary": "Change the destination of a specific parcel delivery order",
        "description": "Change the destination of a specific parcel delivery order. Only the user who created the parcel should be able to change the destination of the parcel. A parcel’s destination can only be changed if it is yet to be delivered.",
        "operationId": "changeOrderDestination",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          }
        },
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Authorization Token",
            "required": false,
            "type": "string"
          },
          {
            "in": "path",
            "name": "parcelId",
            "description": "Parcel Id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Order destination",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderDestination"
            }
          }
        ],
        "security": [
          {
            "token": []
          }
        ]
      }
    },
    "/parcels/{parcelId}/currentlocation": {
      "patch": {
        "tags": [
          "parcel"
        ],
        "summary": "Change the present location of a specific parcel delivery order.",
        "description": "Change the present location of a specific parcel delivery order. Only the Admin is allowed to access this endpoint.",
        "operationId": "changeCurentLocation",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          }
        },
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Token",
            "required": false,
            "type": "string"
          },
          {
            "in": "path",
            "name": "parcelId",
            "description": "Parcel Id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Order Status",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderCurrentlocation"
            }
          }
        ],
        "security": [
          {
            "token": []
          }
        ]
      }
    },
    "/parcels/{parcelId}/status": {
      "patch": {
        "tags": [
          "parcel"
        ],
        "summary": "Change the status of a specific parcel delivery order.",
        "description": "Change the status of a specific parcel delivery order. Only the Admin is allowed to \naccess this endpoint. \n",
        "operationId": "ChangeOrderStatus",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "UnAuthorized"
          }
        },
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Token",
            "required": false,
            "type": "string"
          },
          {
            "name": "parcelId",
            "in": "path",
            "description": "Parcel Id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Order Status",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderStatus"
            }
          }
        ],
        "security": [
          {
            "token": []
          }
        ]
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "othernames": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "registered": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "isadmin": {
          "type": "boolean"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "UserLogin": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "UserSignUp": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "othernames": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "isadmin": {
          "type": "boolean"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "Parcel": {
      "type": "object",
      "required": [
        "placedby",
        "weight",
        "weightmetric",
        "to",
        "from"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "weight": {
          "type": "number"
        },
        "weightmetric": {
          "type": "string",
          "example": "Kg"
        },
        "sentOn": {
          "type": "string"
        },
        "deliveredOn": {
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "currentlocation": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "description": "Parcel status in the store",
          "enum": [
            "placed",
            "transiting",
            "delivered"
          ]
        },
        "isCanceled": {
          "type": "boolean"
        },
        "placedby": {
          "type": "integer",
          "format": "int64"
        }
      },
      "xml": {
        "name": "Parcel"
      }
    },
    "CreateParcel": {
      "type": "object",
      "required": [
        "placedby",
        "weight",
        "weightmetric",
        "to",
        "from"
      ],
      "properties": {
        "weight": {
          "type": "number"
        },
        "weightmetric": {
          "type": "string",
          "example": "Kg"
        },
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "placedby": {
          "type": "integer",
          "format": "int64"
        }
      },
      "xml": {
        "name": "Parcel"
      }
    },
    "OrderDestination": {
      "type": "object",
      "properties": {
        "to": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Parcel Destination"
      }
    },
    "OrderStatus": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "description": "Parcel status in the store",
          "enum": [
            "placed",
            "transiting",
            "delivered"
          ]
        }
      },
      "xml": {
        "name": "Parcel Destination"
      }
    },
    "OrderCurrentlocation": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "currentlocation": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Parcel Current Location"
      }
    }
  },
  "securityDefinitions": {
    "token": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization Token"
    }
  }
}